requirements.txt
gensim
sumy
transformers
torch
torchvision
sentencepiece
nltk
numpy
networkx
rouge
bert_score
bert-extractive-summarizer
pandas